requirements.txt
flask
pillow
flask_cors
requests